refactor: switch commons to bukkit-common
