--- src/main/kotlin/io/github/slaxnetwork/config/model/map/ConfigMapMetaModel.kt
@@ -1,9 +1,0 @@
-package io.github.slaxnetwork.config.model.map
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ConfigMapMetaModel(
-    val name: String,
-    val contributors: List<String>
-)
