+++ src/main/kotlin/io/github/slaxnetwork/config/types/SoundsConfig.kt
@@ -1,6 +1,6 @@
 package io.github.slaxnetwork.config.types
 
-import io.github.slaxnetwork.config.model.ConfigSoundModel
+import io.github.slaxnetwork.bukkitcommon.config.model.ConfigSoundModel
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
